chore:
- reorganized code files
- added docs for breakdown
